release all paa files under APL-SA
https://www.bistudio.com/community/licenses/arma-public-license-share-alike
Attribution - You must attribute the material in the manner specified by
the author or licensor (but not in any way that suggests that they
endorse you or your use of the material).
Noncommercial - You may not use this material for any commercial
purposes.
Arma Only - You may not convert or adapt this material to be used in
other games than Arma.
Share Alike - If you adapt, or build upon this material, you may
distribute the resulting material only under the same license.